Added optional width/height arguments to all prompt types
They have default values of 500 and 200 respectively. Client code can adjust these values as needed. They ensure proper alignment and display of the prompt window (several issues were rooted in the fact that prompts had no dimensions of their own, and had to wait until the inner content rendered and stabilized to get a sense of how big they are).

Also touched on the PromptBaseUI.mxml class, to ensure that buttons stick to the bottom of the window, regardless of text/icon size.